ci(plugin-server): separate out e2e tests from unit tests in ci
