+++ .github/workflows/ci-plugin-server.yml
@@ -134,4 +134,85 @@ jobs:
                   # Below DB name has `test_` prepended, as that's how Django (ran above) creates the test DB
                   DATABASE_URL: 'postgres://posthog:posthog@localhost:5432/test_posthog'
                   REDIS_URL: 'redis://localhost'
-              run: cd plugin-server && yarn test --shard=${{matrix.shard}}
+              run: cd plugin-server && yarn jest tests/ --shard=${{matrix.shard}}
+
+    e2e:
+        name: E2E Tests
+        runs-on: ubuntu-20.04
+
+        env:
+            REDIS_URL: 'redis://localhost'
+            CLICKHOUSE_HOST: 'localhost'
+            CLICKHOUSE_DATABASE: 'posthog_test'
+            KAFKA_HOSTS: 'kafka:9092'
+
+        steps:
+            - name: Code check out
+              uses: actions/checkout@v2
+
+            - name: Stop/Start stack with Docker Compose
+              run: |
+                  docker-compose -f docker-compose.dev.yml down
+                  docker-compose -f docker-compose.dev.yml up -d
+
+            - name: Add Kafka to /etc/hosts
+              run: echo "127.0.0.1 kafka" | sudo tee -a /etc/hosts
+
+            - name: Set up Python
+              uses: actions/setup-python@v2
+              with:
+                  python-version: 3.8.14
+
+            - uses: syphar/restore-virtualenv@v1.2
+              id: cache-backend-tests
+              with:
+                  custom_cache_key_element: v1-
+
+            - uses: syphar/restore-pip-download-cache@v1
+              if: steps.cache-backend-tests.outputs.cache-hit != 'true'
+
+            - name: Install python dependencies
+              if: steps.cache-backend-tests.outputs.cache-hit != 'true'
+              run: |
+                  python -m pip install -r requirements-dev.txt
+                  python -m pip install -r requirements.txt
+
+            - name: Set up Node 16
+              uses: actions/setup-node@v2
+              with:
+                  node-version: 16
+
+            - name: Get yarn cache directory path
+              id: yarn-cache-dir-path
+              run: echo "::set-output name=dir::$(yarn config get cacheFolder)"
+
+            - uses: actions/cache@v2
+              id: yarn-cache
+              with:
+                  path: |
+                      ${{ steps.yarn-cache-dir-path.outputs.dir }}
+                      plugin-server/node_modules
+                  key: plugin-server-yarn-${{ runner.os }}-${{ hashFiles('plugin-server/yarn.lock') }}
+                  restore-keys: |
+                      plugin-server-yarn-${{ runner.os }}
+
+            - name: Install package.json dependencies with Yarn
+              if: steps.yarn-cache.cache-hit != 'true'
+              run: cd plugin-server && yarn
+
+            - name: Wait for Clickhouse & Kafka
+              run: bin/check_kafka_clickhouse_up
+
+            - name: Set up databases
+              env:
+                  TEST: 'true'
+                  SECRET_KEY: 'abcdef' # unsafe - for testing only
+                  DATABASE_URL: 'postgres://posthog:posthog@localhost:5432/posthog'
+              run: cd plugin-server && yarn setup:test
+
+            - name: Test with Jest
+              env:
+                  # Below DB name has `test_` prepended, as that's how Django (ran above) creates the test DB
+                  DATABASE_URL: 'postgres://posthog:posthog@localhost:5432/test_posthog'
+                  REDIS_URL: 'redis://localhost'
+              run: cd plugin-server && yarn jest e2e/ --shard=${{matrix.shard}}
