+++ plugin-server/e2e/e2e.timeout.test.ts
@@ -4,10 +4,10 @@ import { Hub, LogLevel, PluginsServerConfig } from '../src/types'
 import { delay, UUIDT } from '../src/utils/utils'
 import { makePiscina } from '../src/worker/piscina'
 import { createPosthog, DummyPostHog } from '../src/worker/vm/extensions/posthog'
-import { delayUntilEventIngested, resetTestDatabaseClickhouse } from './helpers/clickhouse'
-import { resetKafka } from './helpers/kafka'
-import { pluginConfig39 } from './helpers/plugins'
-import { resetTestDatabase } from './helpers/sql'
+import { delayUntilEventIngested, resetTestDatabaseClickhouse } from '../tests/helpers/clickhouse'
+import { resetKafka } from '../tests/helpers/kafka'
+import { pluginConfig39 } from '../tests/helpers/plugins'
+import { resetTestDatabase } from '../tests/helpers/sql'
 
 jest.setTimeout(60000) // 60 sec timeout
