+++ plugin-server/jest.config.js
@@ -6,6 +6,6 @@ module.exports = {
     clearMocks: true,
     coverageProvider: 'v8',
     setupFilesAfterEnv: ['./jest.setup.fetch-mock.js'],
-    testMatch: ['<rootDir>/tests/**/*.test.ts'],
+    testMatch: ['<rootDir>/**/*.test.ts'],
     testTimeout: 60000,
 }
